requirements.txt
# Core dependencies - versions alignées avec l'environnement de développement
streamlit==1.45.1
pandas==2.2.0
numpy==1.26.4
plotly==5.19.0

# Data processing
openpyxl==3.1.2
xlsxwriter==3.2.0

# Streamlit ecosystem
altair==5.2.0
pydeck==0.8.1b0

#SQL
SQLAlchemy>=2.0
PyMySQL

# Core Python packages
requests==2.31.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1

# Development and testing
pytest==8.0.0
pytest-cov==4.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.12.0

# Utilities
click==8.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
setuptools==69.1.0
six==1.16.0

# Additional packages for your app
blinker==1.7.0
cachetools==5.3.2
attrs==23.2.0
packaging==23.2
typing_extensions==4.9.0

# Git integration
GitPython==3.1.42
gitdb==4.0.11
smmap==5.0.1

# Configuration and data
PyYAML==6.0.1
toml==0.10.2
iniconfig==2.0.0

# Template and markup
Jinja2==3.1.3
MarkupSafe==2.1.5

# JSON schema validation
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
referencing==0.33.0
rpds-py==0.18.0

# Code quality
coverage==7.4.1
pluggy==1.4.0

# Report generation
reportlab==4.1.0

# Async and concurrency
tenacity==8.2.3

# Web framework
tornado==6.4

# Time zone data
tzdata==2024.1

# Data validation
narwhals==0.7.12

# Protocol buffers
protobuf==4.25.3

# Code highlighting
Pygments==2.17.2

# Arrow data format
pyarrow==15.0.0

# Excel file handling
et_xmlfile==1.1.0